# docs/api/openapi-spec.yml
...
openapi: 3.1.0
info:
  title: LogPilot Agent API
  description: 'Cloud-native, autonomous log-intelligence agent: ingestion, search, chat, RCA, reports, proactive
    failure forecasting and autonomy policy.


    Authenticate with POST /api/v1/auth/login, then send `Authorization: Bearer <token>`. Real-time events and streamed
    chat use the WebSocket at /ws?token=<jwt> (not described by OpenAPI).'
  version: 2.0.0
paths:
...
  /api/v1/projects/{project_id}/alerts:
    get:
      tags:
      - alerts
      summary: List Alerts
      operationId: list_alerts_api_v1_projects__project_id__alerts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: open
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alerts/{alert_id}:
    get:
      tags:
      - alerts
      summary: Alert Detail
      operationId: alert_detail_api_v1_projects__project_id__alerts__alert_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/approvals:
    get:
      tags:
      - alerts
      summary: Approvals Queue
      description: 'Everything awaiting a human decision: alerts routed to review by policy + propose-only actions.'
      operationId: approvals_queue_api_v1_projects__project_id__approvals_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/actions/{action_id}/approve:
    post:
      tags:
      - alerts
      summary: Approve Action
      operationId: approve_action_api_v1_projects__project_id__actions__action_id__approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Action Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/actions/{action_id}/edit:
    post:
      tags:
      - alerts
      summary: Edit Action
      operationId: edit_action_api_v1_projects__project_id__actions__action_id__edit_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Action Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__routers__alerts__EditBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/actions/{action_id}/dismiss:
    post:
      tags:
      - alerts
      summary: Dismiss Action
      description: 'Dismiss requires a reason: it is recorded and becomes learning signal.'
      operationId: dismiss_action_api_v1_projects__project_id__actions__action_id__dismiss_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Action Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alerts/{alert_id}/acknowledge:
    post:
      tags:
      - alerts
      summary: Acknowledge
      operationId: acknowledge_api_v1_projects__project_id__alerts__alert_id__acknowledge_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alerts/{alert_id}/dismiss:
    post:
      tags:
      - alerts
      summary: Dismiss Alert
      operationId: dismiss_alert_api_v1_projects__project_id__alerts__alert_id__dismiss_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alerts/{alert_id}/approve:
    post:
      tags:
      - alerts
      summary: Approve Alert
      description: 'An alert that policy routed to human review (pending_review) is released: it becomes a live
        alert and is notified externally.'
      operationId: approve_alert_api_v1_projects__project_id__alerts__alert_id__approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alerts/{alert_id}/outcome:
    post:
      tags:
      - alerts
      summary: Log Outcome
      description: 'After every incident - prevented or not - engineers log whether the alert was accurate, what
        was done and how long

        resolution took. The forecasting subsystem learns from this (weights + leading-indicator signatures).'
      operationId: log_outcome_api_v1_projects__project_id__alerts__alert_id__outcome_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutcomeBody'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alerts/{alert_id}/pre-mortem:
    post:
      tags:
      - alerts
      summary: Pre Mortem
      operationId: pre_mortem_api_v1_projects__project_id__alerts__alert_id__pre_mortem_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OutcomeBody:
      properties:
        outcome:
          type: string
          enum:
          - prevented
          - occurred
          - false_positive
          title: Outcome
        action_taken:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Action Taken
        time_to_resolve_minutes:
          anyOf:
          - type: number
            maximum: 100000.0
            minimum: 0.0
          - type: 'null'
          title: Time To Resolve Minutes
        notes:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Notes
      type: object
      required:
      - outcome
      title: OutcomeBody
...
    ReasonBody:
      properties:
        reason:
          type: string
          maxLength: 2000
          minLength: 3
          title: Reason
      type: object
      required:
      - reason
      title: ReasonBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    app__routers__alerts__EditBody:
      properties:
        text:
          type: string
          maxLength: 2000
          minLength: 3
          title: Text
        approve:
          type: boolean
          title: Approve
          default: true
      type: object
      required:
      - text
      title: EditBody
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local (docker compose)